requirements.txt
streamlit
pypdf
python-docx
langchain
sentence-transformers
faiss-cpu
transformers
llama-cpp-python